write the Bayesian layer properly, and position against the right baseline
The method section opened with a metaphor and treated the probabilistic content
as an implementation note.  It now opens where it belongs: neural functional
tensor decomposition already replaced the discrete factor with a function of
position, and graph-Laplacian regularisation already brings connectivity in as a
penalty; this work puts connectivity into the parameterisation instead.  A table
sets the four approaches side by side, and the paragraph that matters says what
happens to an unobserved node under each -- unconstrained, constrained through
coordinate smoothness, constrained through graph neighbours, or determined by
the same K x R coefficients as every other node.

The Bayesian treatment is now derived rather than asserted.  A notation table, a
full generative model, and the observation that carries everything -- given the
factors the model is linear in the core, which is why an exact conditional
posterior exists.  The spectral prior is written as a Gaussian whose precision
is (1+lambda_k)^p, making explicit that the geometry supplies the dictionary and
the prior at once rather than a dictionary plus a smoothness penalty bolted on.
Then the evidence, its log, the MacKay fixed point, and the effective number of
parameters in the form that shows what it means: a sum of per-direction terms
saying how far each direction is determined by data rather than by the prior.
The predictive distribution now says why a calibration factor is needed at all
-- the factors are point estimates, so the core posterior alone underestimates.

The pre-fit diagnostics move out of the method and are labelled as what they
are: our own tool for deciding whether a candidate dataset is worth GPU time,
explicitly not paper material, with a use case named for each.

The graph-Laplacian baseline is also back in the runner, which it had not been
since the sparse rewrite -- the dense penalty needed three N-by-N
eigendecompositions.  As M + K/pi^2 it is exactly the Sobolev price at p = 1,
which is the standard graph regulariser, and it is sparse.  Early rows already
show the control behaving: on the barrier-free layout the two Laplacian variants
return identical numbers, as the two spectral variants do.

Report is 18 pages, zero overfull boxes.  64 tests pass.

+++ src/geoaware/grouped_operator_tucker.py
@@ -196,7 +196,9 @@ class GroupedOperatorTucker(nn.Module):
                 if spec.eigenvalues is not None:
                     spec.eigenvalues = spec.eigenvalues.float().to(device)
                 if spec.penalty_operator is not None:
-                    spec.penalty_operator = spec.penalty_operator.float().to(device)
+                    # Sparse penalties stay sparse: the dense form of the same
+                    # matrix does not fit at these mesh sizes.
+                    spec.penalty_operator = spec.penalty_operator.to(device)
             self._buffers_moved = True
         return device
